Changed name of Gamerunner to TESTMAIN Changed name of modes to PersistenceManager Removed inheritance in modes Added txt file for Persistence called UserData
